refactor : frontend http mode 타입 추가 및 정리
