+++ frontend/app/utils/denchUitl.ts
@@ -1,5 +1,5 @@
-import type { DenchConfig, HTTPMode } from "~/types/utils/simpleFetcher/dench";
-import type { HTTPCredentials } from "~/types/utils/simpleFetcher/denchEnum";
+import type { DenchConfig } from "~/types/utils/simpleFetcher/dench";
+import type { HTTPCredentials, HTTPMode } from "~/types/utils/simpleFetcher/denchEnum";
 import denchfetcher from "./denchfetcher";
 
 export function runfetch<T>(config : DenchConfig) : Promise<Response>{
